requirements.txt
fastapi
uvicorn
streamlit
openai
faiss-cpu
numpy
pandas
python-dotenv
requests
pydantic
python-multipart
sentence-transformers
scikit-learn
